Moved locales out of middleware, fixed tests
