+++ app/[locale]/feed/layout.tsx
@@ -2,7 +2,7 @@ import { Suspense } from 'react';
 
 const FeedLayout = ({ children }: { children: React.ReactNode }) => {
   return (
-    <div className="container mx-auto text-white">
+    <div className="container mx-auto text-white" data-testid="feed-layout">
       <Suspense>{children}</Suspense>
       <div className="h-32"></div>
     </div>
