+++ components/Post/PostHeader.tsx
@@ -5,7 +5,7 @@ import Link from '@/components/Link/Link';
 import LinkButtonNoNumber from '../Buttons/LinkButton/LinkButtonNoNumber';
 import ImageOptimized from '../ImageOptimized/ImageOptimized';
 import { useLocale } from 'next-intl';
-import { localeToDateLocale } from '@/middleware';
+import { localeToDateLocale } from '@/translations/locales';
 
 const PostHeader = ({
   author,
