+++ translations/locales.ts
@@ -1,0 +1,7 @@
+export const locales = ['en', 'de'];
+
+// Used for `toLocaleString` internationalization
+export const localeToDateLocale = {
+  en: 'en-US',
+  de: 'de-DE',
+};
